+++ main/src/main/java/com/proyecto/SueldosMapping.java
@@ -2,7 +2,7 @@ package com.proyecto;
 
 public class SueldosMapping {
     
-    public Sueldos mapSueldos(int SueldoBrutoBD, float SueldoNetoBD){
+    public Sueldos mapSueldos(int SueldoBrutoBD, int SueldoNetoBD){
         Sueldos sueldos = new Sueldos(SueldoBrutoBD, SueldoNetoBD);
         return sueldos;
     }
